# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tender API",
    "version": "0.1.0",
    "description": "Any-chain crypto checkout. Buyers pay from any chain; merchants settle in one asset on Monad.\n\nMerchant routes need **Authorize → `tk_live_…`** (create one with `npm run merchant:create`). Public routes are what the buyer checkout calls. Money is always a decimal **string**; the wire is **snake_case**."
  },
  "components": {
    "securitySchemes": {
      "merchantKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Merchant API key (tk_live_…)"
      },
      "platformKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Dashboard platform key (tp_…). On /v1 routes, also send `X-Tender-Merchant: mer_…`. Server-side only."
      }
    }
  },
...
  "paths": {
...
    "/v1/payments": {
      "get": {
        "tags": [
          "Payments"
        ],
        "summary": "List payments across invoices",
        "security": [
          {
            "merchantKey": []
          },
          {
            "platformKey": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "DETECTED",
                "SETTLED",
                "FAILED",
                "REFUNDED"
              ]
            }
          },
          {
            "name": "invoice_status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "DETECTED",
                "SETTLED",
                "OVERPAID",
                "UNDERPAID",
                "EXPIRED",
                "CANCELLED",
                "NEEDS_RECOVERY"
              ]
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "invoice_id": {
                            "type": "string"
                          },
                          "tx_hash": {
                            "type": "string"
                          },
                          "from_chain": {
                            "type": "string",
                            "minLength": 1
                          },
                          "amount_in": {
                            "type": "string",
                            "pattern": "^(0|[1-9]\\d*)(\\.\\d{1,18})?$"
                          },
                          "amount_settled": {
                            "anyOf": [
                              {
                                "type": "string",
                                "pattern": "^(0|[1-9]\\d*)(\\.\\d{1,18})?$"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "sender": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "refunded_amount": {
                            "anyOf": [
                              {
                                "type": "string",
                                "pattern": "^(0|[1-9]\\d*)(\\.\\d{1,18})?$"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "DETECTED",
                              "SETTLED",
                              "FAILED",
                              "REFUNDED"
                            ]
                          },
                          "first_seen_at": {
                            "type": "string"
                          },
                          "settled_at": {
                            "type": [
                              "string",
                              "null"
                            ]
                          }
                        },
                        "required": [
                          "id",
                          "invoice_id",
                          "tx_hash",
                          "from_chain",
                          "amount_in",
                          "status",
                          "first_seen_at"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "next_cursor": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "has_more": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "data",
                    "has_more"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Validation failed — `fields` names each bad field",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/payments/{id}": {
      "get": {
        "tags": [
          "Payments"
        ],
        "summary": "One payment, with history and any recovery task",
        "security": [
          {
            "merchantKey": []
          },
          {
            "platformKey": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "invoice_id": {
                      "type": "string"
                    },
                    "tx_hash": {
                      "type": "string"
                    },
                    "from_chain": {
                      "type": "string",
                      "minLength": 1
                    },
                    "amount_in": {
                      "type": "string",
                      "pattern": "^(0|[1-9]\\d*)(\\.\\d{1,18})?$"
                    },
                    "amount_settled": {
                      "anyOf": [
                        {
                          "type": "string",
                          "pattern": "^(0|[1-9]\\d*)(\\.\\d{1,18})?$"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "sender": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "refunded_amount": {
                      "anyOf": [
                        {
                          "type": "string",
                          "pattern": "^(0|[1-9]\\d*)(\\.\\d{1,18})?$"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "DETECTED",
                        "SETTLED",
                        "FAILED",
                        "REFUNDED"
                      ]
                    },
                    "first_seen_at": {
                      "type": "string"
                    },
                    "settled_at": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "invoice": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "reference": {
                          "type": "string"
                        },
                        "amount_expected": {
                          "type": "string",
                          "pattern": "^(0|[1-9]\\d*)(\\.\\d{1,18})?$"
                        },
                        "currency": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "PENDING",
                            "DETECTED",
                            "SETTLED",
                            "OVERPAID",
                            "UNDERPAID",
                            "EXPIRED",
                            "CANCELLED",
                            "NEEDS_RECOVERY"
                          ]
                        }
                      },
                      "required": [
                        "id",
                        "reference",
                        "amount_expected",
                        "currency",
                        "status"
                      ],
                      "additionalProperties": false
                    },
                    "history": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "status": {
                            "type": "string"
                          },
                          "at": {
                            "type": "string"
                          },
                          "note": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "status",
                          "at"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "recovery": {
                      "anyOf": [
                        {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "payment_id": {
                              "type": "string"
                            },
                            "reason": {
                              "type": "string"
                            },
                            "state": {
                              "type": "string",
                              "enum": [
                                "OPEN",
                                "RETRYING",
                                "WITHDRAWN",
                                "RESOLVED",
                                "FAILED"
                              ]
                            },
                            "notes": {
                              "type": [
                                "string",
                                "null"
                              ]
                            },
                            "created_at": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "id",
                            "payment_id",
                            "reason",
                            "state",
                            "created_at"
                          ],
                          "additionalProperties": false
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "id",
                    "invoice_id",
                    "tx_hash",
                    "from_chain",
                    "amount_in",
                    "status",
                    "first_seen_at",
                    "invoice",
                    "history"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Not found (or not yours)",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/payments/{id}/refund": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Refund (use POST /v1/transfers instead)",
        "security": [
          {
            "merchantKey": []
          },
          {
            "platformKey": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Not found (or not yours)",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "501": {
            "description": "Always 501: persistent addresses have no refund API"
          }
        }
      }
    },
...
  }
}